0174 Added AuthService import at line 6. Added Private AuthService at line 16. Added isAuthenticated boolean at line 22 for authService. Changed classes to interfaces at line 4. Should we change line 27 to getSignOut?

+++ src/app/shared/components/main-nav.component.ts
@@ -1,8 +1,9 @@
 import {Component} from "@angular/core";
 import {SignInService} from "../services/sign.in.service";
 import {CookieService} from "ngx-cookie-service";
-//import {Status} from "../classes/status";
+// import {Status} from "../../interfaces/status";
 import {Router} from "@angular/router";
+import {AuthService} from "../services/auth.service";
 
 @Component({
 	template: require("./main-nav.component.html"),
@@ -12,26 +13,30 @@ import {Router} from "@angular/router";
 export class MainNavComponent {
 
 	constructor(
+		private authService: AuthService,
 		private signInService: SignInService,
 		private cookieService: CookieService,
 		private router: Router
 	) {}
 
-
+	isAuthenticated(): boolean {
+		return(this.authService.isAuthenticated());
+	}
 
 	signOut() : void {
-		this.signInService.signOut();
+		this.signInService.signOut()
+		// should this be getSignOut?
 
-//			.subscribe(status => {
-//				this.status = status;
-//				if(status.status === 200) {
+			.subscribe(status => {
+				this.status = status;
+				if(status.status === 200) {
 
 					//delete cookies and jwt
 					this.cookieService.deleteAll();
 					localStorage.clear();
 
-					//send user back home, refresh page
-					this.router.navigate([""]);
+					//send user back to home view, refresh page
+					this.router.navigate(["signed-out"]);
 					location.reload();
 					console.log("goodbye");
 				}
